created SpotifyBaseException and moved exceptions from oauth2.py to exceptions.py (#1161)
